requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
manim>=0.19.0
google-cloud-storage>=2.10.0
redis>=5.0.1
pillow>=10.1.0
anthropic>=0.18.1
python-multipart>=0.0.6
